improved index page legend
